Added important main features of the program
Added a function which shows information of user who is currently logged in; added a function which checks if new record's ID already exists in the file. This removes ID duplication. Another function is added which allows Teachers and Students to update their record. Only a few selected fields are allowed to be updated. Sorting of record has been started